move shared code to astria-sequencer-validation crate
