+++ Cargo.toml
@@ -8,6 +8,8 @@ members = [
   "crates/astria-sequencer",
   "crates/astria-sequencer-client",
   "crates/astria-sequencer-relayer",
+  "crates/astria-sequencer-validation",
+  "crates/astria-sequencer-types",
   "crates/astria-sequencer-utils",
   "crates/astria-test-utils",
 ]
