+++ crates/astria-sequencer-relayer/Cargo.toml
@@ -41,6 +41,7 @@ astria-gossipnet = { path = "../astria-gossipnet" }
 astria-proto = { path = "../astria-proto" }
 astria-sequencer = { path = "../astria-sequencer" }
 astria-sequencer-types = { path = "../astria-sequencer-types" }
+astria-sequencer-validation = { path = "../astria-sequencer-validation" }
 
 [dev-dependencies]
 astria-celestia-jsonrpc-client = { path = "../astria-celestia-jsonrpc-client", features = [
