+++ crates/astria-sequencer-validation/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "astria-sequencer-validation"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1"
+ct-merkle = "0.1"
+
+bytes = { workspace = true }
+hex = { workspace = true, features = ["serde"] }
+serde = { workspace = true, features = ["derive"] }
+sha2 = { workspace = true }
+tendermint = { workspace = true }
+
+[dev-dependencies]
+serde_json = { workspace = true }
