+++ crates/astria-sequencer-validation/src/lib.rs
@@ -1,0 +1,8 @@
+mod proof;
+mod utils;
+
+pub use proof::{
+    InclusionProof,
+    MerkleTree,
+};
+pub use utils::generate_action_tree_leaves;
