+++ crates/astria-sequencer-validation/src/proof.rs
