+++ crates/astria-sequencer/src/proposal/commitment.rs
@@ -1,5 +1,6 @@
 use std::collections::BTreeMap;
 
+use astria_sequencer_validation::generate_action_tree_leaves;
 use bytes::Bytes;
 use tendermint::merkle::simple_hash_from_byte_vectors;
 
@@ -48,23 +49,6 @@ pub(crate) fn generate_sequence_actions_commitment(
         simple_hash_from_byte_vectors::<tendermint::crypto::default::Sha256>(&leaves),
         txs_to_include,
     )
-}
-
-/// Groups the `sequence::Action`s within the transactions by their `chain_id`.
-/// The `BTreeMap` key is the chain ID and value is the transactions.
-#[must_use]
-pub fn generate_action_tree_leaves(
-    chain_id_to_txs: &BTreeMap<Vec<u8>, Vec<Vec<u8>>>,
-) -> Vec<Vec<u8>> {
-    let mut leaves: Vec<Vec<u8>> = vec![];
-    for (chain_id, txs) in chain_id_to_txs {
-        let chain_id_root =
-            simple_hash_from_byte_vectors::<tendermint::crypto::default::Sha256>(txs);
-        let mut leaf = chain_id.clone();
-        leaf.append(&mut chain_id_root.to_vec());
-        leaves.push(leaf);
-    }
-    leaves
 }
 
 /// Groups the `sequence::Action`s within the transactions by their `chain_id`.
