+++ crates/astria-sequencer/src/proposal/mod.rs
@@ -1,2 +1,1 @@
 pub mod commitment;
-pub mod proof;
